fix: split facet-profile out of facet-interact to satisfy route-gallery contract
Contract verifiers' 'route gallery contract' (Rule 5) failed on the prior
commit: facet-interact.vue grew from 204 to 228 lines by adding the new
Storybook CTA inline, but the interact tier is a router, not a working
surface -- model-specific content belongs in a dedicated activity
component (bot-chat, dream-narration, reward-encounter, scenario-story).
Facet had no such split yet.

Extracted the whole detail body (header, art plate, description, badges,
the #detail slot, and the new 'Start a story with this' button) into a
new facet-profile.vue, mirroring dream-interact -> dream-narration and
reward-interact -> reward-encounter. facet-interact.vue is back to being
a thin gallery/profile router (98 lines, down from 204 -- well under the
route-gallery-baseline.json ratchet).

Verified: eslint clean, prettier --check clean, vue-tsc --noEmit exit 0,
test:route-gallery-contract / test:facet-gallery / test:gallery-adoption /
test:gallery-consistency / test:lint-ratchet / test:component-reachability
all pass (346 live components, 0 unreachable).
