+++ components/facets/facet-interact.vue
@@ -1,23 +1,14 @@
 <!-- /components/facets/facet-interact.vue -->
 <!--
-  The Facet working surface, and the tier Facets was missing.
+  Router for the Facets working surface: gallery until something is picked,
+  then facet-profile. Mirrors dream-interact -> dream-narration and
+  reward-interact -> reward-encounter -- the frame is shared here, the
+  profile content lives in its own file so this stays a router rather than
+  the working surface (Rule 5, utils/scripts/verifyRouteGalleryContract.ts).
 
-  Every other model follows the same shape -- content/<model>.md mounts
-  :<model>-manager, which routes to <model>-interact, which shows a gallery until
-  you pick something and a surface surface once you have:
-
-    bot-interact    <bot-gallery   v-if="!botStore.currentBot">   ... else the chat
-    dream-interact  <dream-gallery v-if="!selectedDream">  <dream-narration v-else>
-
-  Facets had neither tier. facet-manager rendered all ~1611 rows itself and
-  expanded an editor INSIDE the grid cell, which is why selecting one felt wrong
-  and why /facets scored 3/10 on review while Dreams, which follows the pattern,
-  scored 7.5.
-
-  Selection lives in the URL rather than in local state, per the house rule that
-  every view is linkable (the same shape as /storybook?story=...). That also
-  means the browser back button does the obvious thing, which was the other half
-  of what made the inline editor feel wrong.
+  Selection lives in the URL rather than in local state, per the house rule
+  that every view is linkable (the same shape as /storybook?story=...). That
+  also means the browser back button does the obvious thing.
 -->
 <template>
   <div v-if="!selectedFacet" class="flex h-full min-h-0 flex-col gap-2">
@@ -40,98 +31,11 @@
     />
   </div>
 
-  <section v-else class="kr-surface">
-    <header class="kr-toolbar shrink-0 kr-panel-flat p-3">
-      <button
-        type="button"
-        class="btn btn-ghost btn-sm rounded-xl"
-        @click="closeFacet"
-      >
-        <Icon name="kind-icon:chevron-left" class="size-4" /> All Facets
-      </button>
-
-      <div class="min-w-0 flex-1">
-        <p class="truncate font-black">{{ selectedFacet.title }}</p>
-        <p class="truncate text-xs text-base-content/55">
-          {{ taxonomyLabel(selectedFacet.taxonomy) }}
-          <template v-if="selectedFacet.groupLabel">
-            · {{ selectedFacet.groupLabel }}
-          </template>
-        </p>
-      </div>
-
-      <span class="badge badge-outline badge-sm shrink-0">
-        {{ selectedFacet.canonicalValue }}
-      </span>
-
-      <button
-        type="button"
-        class="btn btn-primary btn-sm shrink-0 rounded-xl"
-        title="Start a Storybook story seeded with this Facet"
-        @click="startStoryWithFacet"
-      >
-        <Icon name="kind-icon:book-open" class="size-3.5" />
-        <span class="hidden sm:inline">Start a story with this</span>
-      </button>
-    </header>
-
-    <div class="kr-scroll space-y-4">
-      <div class="grid gap-4 lg:grid-cols-[minmax(0,22rem)_minmax(0,1fr)]">
-        <kr-art-plate
-          :source="selectedFacet"
-          variant="card"
-          shape="card"
-          :alt="`${selectedFacet.title} artwork`"
-          :placeholder-icon="iconName(selectedFacet)"
-        />
-
-        <div class="space-y-3">
-          <p v-if="selectedFacet.description" class="kr-prose leading-relaxed">
-            {{ selectedFacet.description }}
-          </p>
-          <p
-            v-if="selectedFacet.flavorText"
-            class="kr-prose italic text-base-content/65"
-          >
-            {{ selectedFacet.flavorText }}
-          </p>
-
-          <div
-            v-if="selectedFacet.aliases.length"
-            class="flex flex-wrap gap-1.5"
-          >
-            <span
-              v-for="alias in selectedFacet.aliases"
-              :key="alias"
-              class="badge badge-ghost badge-sm"
-            >
-              {{ alias }}
-            </span>
-          </div>
-
-          <div class="flex flex-wrap gap-1.5 text-xs">
-            <span class="badge badge-sm"
-              >order {{ selectedFacet.sortOrder }}</span
-            >
-            <span class="badge badge-sm"
-              >weight {{ selectedFacet.randomWeight }}</span
-            >
-            <span v-if="selectedFacet.isRandomizable" class="badge badge-sm">
-              randomizable
-            </span>
-            <span
-              v-if="selectedFacet.artRequired"
-              class="badge badge-warning badge-sm"
-            >
-              art expected
-            </span>
-          </div>
-        </div>
-      </div>
-
-      <slot name="detail" :facet="selectedFacet" />
-    </div>
-  </section>
+  <facet-profile v-else>
+    <template #detail="slotProps">
+      <slot name="detail" v-bind="slotProps" />
+    </template>
+  </facet-profile>
 </template>
 
 <script setup lang="ts">
@@ -164,12 +68,6 @@ function openFacet(facet: FacetCatalogEntry): void {
   void router.push({ query: { ...route.query, facet: facet.slug } })
 }
 
-function closeFacet(): void {
-  const query = { ...route.query }
-  delete query.facet
-  void router.push({ query })
-}
-
 /*
  * The gallery is a deliberately read-only surface (verifyFacetGallery.ts
  * forbids create/update/archive here) -- creation belongs to the Library
@@ -187,39 +85,11 @@ function goToCreateFacet(): void {
 }
 
 /*
- * Facets are a tool that feeds other interactions rather than something a
- * visitor talks to directly (Silas: Facets and Rewards should link out to
- * Storybook instead of growing their own conversation). storybook-page.vue's
- * seedFromQuery() reads ?facet=<slug> and adds it to the setup draft's
- * facetSlugs on top of whatever draft is already restored, so this never
- * destroys a story in progress.
- */
-function startStoryWithFacet(): void {
-  if (!selectedFacet.value?.slug) return
-  void navigateTo({
-    path: '/storybook',
-    query: { facet: selectedFacet.value.slug },
-  })
-}
-
-function taxonomyLabel(taxonomy: string): string {
-  return taxonomy
-    .toLowerCase()
-    .replace(/_/g, ' ')
-    .replace(/\b\w/g, (letter) => letter.toUpperCase())
-}
-
-function iconName(facet: FacetCatalogEntry): string {
-  const icon = facet.icon?.trim()
-  return icon && icon.includes(':') ? icon : 'kind-icon:tag'
-}
-
-/*
  * The gallery fetches the catalog on mount, but a deep link to ?facet=<slug>
- * renders the detail branch instead and would otherwise resolve against an empty
- * store -- landing the visitor back on the gallery for a link that was perfectly
- * valid. fetchCatalog no-ops once loaded, so this costs nothing on the normal
- * browse-then-select path.
+ * renders the profile branch instead and would otherwise resolve against an
+ * empty store -- landing the visitor back on the gallery for a link that was
+ * perfectly valid. fetchCatalog no-ops once loaded, so this costs nothing on
+ * the normal browse-then-select path.
  */
 onMounted(() => {
   if (!catalog.loaded) void catalog.fetchCatalog()
